models/SkeletonActionRecognition/requirements.txt
about-time==4.2.1
accelerate==1.3.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.2
alive-progress==3.2.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
autograd==1.7.0
babel==2.16.0
beautifulsoup4==4.12.3
black==25.1.0
bleach==6.2.0
cap_from_youtube==0.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
cma==3.2.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fake-useragent==2.0.3
fastapi==0.115.8
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.17.0
flake8==7.1.1
flatbuffers==25.1.24
fonttools==4.55.4
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
ghp-import==2.1.0
gradio==4.29.0
gradio_client==0.16.1
gradio_imageslider==0.0.20
grapheme==0.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
imread_from_url==0.1.3
iniconfig==2.0.0
interrogate==1.7.0
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==4.3.21
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json-tricks==3.17.3
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jstyleson==0.0.2
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mat4py==0.6.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.0
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
model-index==0.1.11
modelindex==0.0.2
motmetrics==1.4.0
mpmath==1.3.0
multidict==6.2.0
munkres==1.1.4
mypy-extensions==1.0.0
narwhals==1.26.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
nncf==2.14.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.11.0.86
openvino==2024.6.0
openvino-telemetry==2025.0.0
ordered-set==4.1.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
poseval @ git+https://github.com/svenkreiss/poseval.git@3128c5cbcf90946e5164ff438ad651e113e64613
prettytable==3.12.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.21.6
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==2.0.0
pydub==0.25.1
pyflakes==3.2.0
pyglet==2.1.2
Pygments==2.19.1
pymoo==0.6.1.3
PyOpenGL==3.1.0
pyparsing==3.2.1
pyrender==0.1.45
pytest==8.3.4
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.2.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.4
safehttpx==0.1.6
safetensors==0.5.2
SAM-2 @ git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
sam2==1.1.0
sapiens-inferece==0.2.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
semantic-version==2.10.0
Send2Trash==1.8.3
shapely==2.0.6
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smplx==0.1.28
sniffio==1.3.1
soupsieve==2.6
spaces==0.32.0
stack-data==0.6.3
starlette==0.45.3
stringzilla==3.11.3
supervision==0.25.1
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.5.0
timm==1.0.13
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.12.0
torch~=2.6.0
torch-geometric==2.6.1
torchaudio==2.5.1
torchsummary==1.5.1
torchvision~=0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@222505c7e4d08da9095d12ddb72fb653f4b6da33
trimesh==4.6.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.59
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xdoctest==1.2.0
xmltodict==0.14.2
xtcocotools==1.14.3
yapf==0.40.1
yarl==1.18.3
youtube-dl==2021.12.17
yt-dlp==2025.1.15
zipp==3.21.0

prefect~=3.2.14
transformers~=4.50.0.dev0